workshop_projects/projects/crewai/simple_data_analysis_crew/requirements.txt
crewai
streamlit
pandas
python-dotenv
